updated repo location of view post
