+++ src/components/views/styles/view-post.styles.css
@@ -1,0 +1,8 @@
+.view-post-container {
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 100%;
+    height: 100vh;
+}
